move syste/glut.h into the glut project
