+++ applications/projects/SofaGuiGlut/SimpleGUI.h
@@ -24,6 +24,8 @@
 
 #include <sofa/gui/BaseGUI.h>
 
+#include "glut.h"
+
 #include <sofa/gui/PickHandler.h>
 
 #include <sofa/helper/system/config.h>
@@ -34,7 +36,6 @@
 #include <sofa/helper/system/thread/CTime.h>
 #include <sofa/helper/system/gl.h>
 #include <sofa/helper/system/glu.h>
-#include <sofa/helper/system/glut.h>
 #include <sofa/core/visual/VisualParams.h>
 #include <sofa/core/visual/DrawToolGL.h>
 #include <SofaBaseVisual/InteractiveCamera.h>
